+++ src/torchmetrics/utilities/compute.py
@@ -11,8 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from typing import Tuple
+
 import torch
 from torch import Tensor
+
+from torchmetrics.utilities.imports import _TORCH_GREATER_EQUAL_1_9
 
 
 def _safe_matmul(x: Tensor, y: Tensor) -> Tensor:
@@ -49,3 +53,63 @@ def _safe_divide(num: Tensor, denom: Tensor) -> Tensor:
     num = num if num.is_floating_point() else num.float()
     denom = denom if denom.is_floating_point() else denom.float()
     return num / denom
+
+
+def _auc_update(x: Tensor, y: Tensor) -> Tuple[Tensor, Tensor]:
+    """Checks that auc input is correct."""
+    x = x.squeeze() if x.ndim > 1 else x
+    y = y.squeeze() if y.ndim > 1 else y
+
+    if x.ndim > 1 or y.ndim > 1:
+        raise ValueError(
+            f"Expected both `x` and `y` tensor to be 1d, but got tensors with dimension {x.ndim} and {y.ndim}"
+        )
+    if x.numel() != y.numel():
+        raise ValueError(
+            f"Expected the same number of elements in `x` and `y` tensor but received {x.numel()} and {y.numel()}"
+        )
+    return x, y
+
+
+def _auc_compute_without_check(x: Tensor, y: Tensor, direction: float) -> Tensor:
+    """Computes area under the curve using the trapezoidal rule.
+
+    Assumes increasing or decreasing order of `x`.
+    """
+    with torch.no_grad():
+        auc_: Tensor = torch.trapz(y, x) * direction
+    return auc_
+
+
+def _auc_compute(x: Tensor, y: Tensor, reorder: bool = False) -> Tensor:
+    with torch.no_grad():
+        if reorder:
+            x, x_idx = torch.sort(x, stable=True) if _TORCH_GREATER_EQUAL_1_9 else torch.sort(x)
+            y = y[x_idx]
+
+        dx = x[1:] - x[:-1]
+        if (dx < 0).any():
+            if (dx <= 0).all():
+                direction = -1.0
+            else:
+                raise ValueError(
+                    "The `x` tensor is neither increasing or decreasing. Try setting the reorder argument to `True`."
+                )
+        else:
+            direction = 1.0
+        return _auc_compute_without_check(x, y, direction)
+
+
+def auc(x: Tensor, y: Tensor, reorder: bool = False) -> Tensor:
+    """Computes Area Under the Curve (AUC) using the trapezoidal rule.
+
+    Args:
+        x: x-coordinates, must be either increasing or decreasing
+        y: y-coordinates
+        reorder: if True, will reorder the arrays to make it either increasing or decreasing
+
+    Return:
+        Tensor containing AUC score
+    """
+    x, y = _auc_update(x, y)
+    return _auc_compute(x, y, reorder=reorder)
